// prodis-prodiswcl/src/app/modules/prodisapi/model/statoConcessione.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProvCompensazioni } from './provCompensazioni';
import { ProvConvenzione } from './provConvenzione';
import { ProvEsonero } from './provEsonero';
import { ProvSospensione } from './provSospensione';


export interface StatoConcessione { 
    codStatoConcessione?: string;
    dataFine?: Date;
    dataInizio?: Date;
    dataTmst?: Date;
    descStatoConcessione?: string;
    id?: number;
    provCompensazionis?: Array<ProvCompensazioni>;
    provConvenziones?: Array<ProvConvenzione>;
    provEsoneros?: Array<ProvEsonero>;
    provSospensiones?: Array<ProvSospensione>;
}